Добавляет Smart компонент
